refactor(toolchain): extract some types from parser code so they can be shared
