+++ sirc-vm/toolchain/src/lib.rs
@@ -17,4 +17,5 @@
 
 pub mod data;
 pub mod parsers;
+pub mod printers;
 pub mod types;
