+++ sirc-vm/toolchain/src/parsers/data.rs
@@ -3,22 +3,44 @@ use nom::{
     combinator::map,
     error::{ErrorKind, FromExternalError},
 };
-use nom_supreme::{error::ErrorTree, tag::complete::tag, ParserExt};
+use nom_supreme::error::ErrorTree;
+use nom_supreme::tag::complete::tag;
+use nom_supreme::ParserExt;
 use serde::Serialize;
 
+use crate::types::{
+    data::{DataToken, DataType, EquToken},
+    object::RefType,
+};
+
 use super::{
-    instruction::RefToken,
+    instruction::Token,
     shared::{lexeme, parse_number, parse_placeholder, parse_symbol_reference, AsmResult},
 };
 
-#[derive(Debug, Serialize)]
-pub enum DataType {
-    Value(u32),
-    SymbolRef(RefToken),
-    PlaceHolder(String),
+#[derive(Debug, Clone, Serialize)]
+pub struct RefToken {
+    pub name: String,
+    pub ref_type: RefType,
 }
 
-pub fn parse_data_type(i: &str) -> AsmResult<DataType> {
+pub fn override_ref_token_type_if_implied(
+    ref_token: &RefToken,
+    override_ref_type: RefType,
+) -> RefToken {
+    match ref_token.ref_type {
+        RefType::Implied => RefToken {
+            name: ref_token.name.clone(),
+            ref_type: override_ref_type,
+        },
+        _ => RefToken {
+            name: ref_token.name.clone(),
+            ref_type: ref_token.ref_type,
+        },
+    }
+}
+
+fn parse_data_type(i: &str) -> AsmResult<DataType> {
     alt((
         // TODO: Make sure that numbers are defined consistently in SIRC Asm
         // category=Toolchain
@@ -31,7 +53,7 @@ pub fn parse_data_type(i: &str) -> AsmResult<DataType> {
     ))(i)
 }
 
-pub fn parse_data_(i: &str) -> AsmResult<(u8, DataType)> {
+fn parse_data_(i: &str) -> AsmResult<(u8, DataType)> {
     let (i, tag) = lexeme(alt((tag(".DB"), tag(".DW"), tag(".DQ"))))(i)?;
     let (i, value) = parse_data_type(i)?;
 
@@ -52,11 +74,11 @@ pub fn parse_data_(i: &str) -> AsmResult<(u8, DataType)> {
     Ok((i, (size, value)))
 }
 
-pub fn parse_data(i: &str) -> AsmResult<(u8, DataType)> {
+fn parse_data(i: &str) -> AsmResult<(u8, DataType)> {
     lexeme(parse_data_)(i)
 }
 
-pub fn parse_equ_(i: &str) -> AsmResult<(String, u32)> {
+fn parse_equ_(i: &str) -> AsmResult<(String, u32)> {
     let (i, _) = lexeme(tag(".EQU"))(i)?;
     let (i, placeholder_name) = parse_placeholder(i)?;
     let (i, value) = parse_number(i)?;
@@ -64,6 +86,39 @@ pub fn parse_equ_(i: &str) -> AsmResult<(String, u32)> {
     Ok((i, (placeholder_name, value)))
 }
 
-pub fn parse_equ(i: &str) -> AsmResult<(String, u32)> {
+fn parse_equ(i: &str) -> AsmResult<(String, u32)> {
     lexeme(parse_equ_)(i)
+}
+
+pub fn parse_data_token(i: &str) -> AsmResult<Token> {
+    let (i, (size_bytes, value)) = parse_data(i)?;
+
+    let override_value = if let DataType::SymbolRef(ref_token) = value {
+        DataType::SymbolRef(override_ref_token_type_if_implied(
+            &ref_token,
+            RefType::FullAddress,
+        ))
+    } else {
+        value
+    };
+
+    Ok((
+        i,
+        Token::Data(DataToken {
+            size_bytes,
+            value: override_value,
+        }),
+    ))
+}
+
+pub fn parse_equ_token(i: &str) -> AsmResult<Token> {
+    let (i, (placeholder_name, value)) = parse_equ(i)?;
+
+    Ok((
+        i,
+        Token::Equ(EquToken {
+            placeholder_name,
+            value,
+        }),
+    ))
 }
