+++ sirc-vm/toolchain/src/printers/data.rs
@@ -1,0 +1,9 @@
+use crate::types::data::{DataToken, EquToken};
+
+pub fn print_data_token(_: DataToken) -> String {
+    return "TODO".to_string();
+}
+
+pub fn print_equ_token(_: EquToken) -> String {
+    return "TODO".to_string();
+}
