+++ sirc-vm/toolchain/src/types/mod.rs
@@ -1,1 +1,2 @@
+pub mod data;
 pub mod object;
